// src/types/IPreset.gen.ts
// Generated by ts-to-zod
import { z } from "zod";

export const presetStepTypeSchema = z.union([
  z.literal("commandline"),
  z.literal("hydrate"),
  z.literal("setgame"),
  z.literal("installmod"),
  z.literal("restart"),
  z.literal("login_nexus"),
]);

export const iPresetStepBaseSchema = z.object({
  type: presetStepTypeSchema,
  id: z.string(),
});

export const iCommandLineArgSchema = z.object({
  key: z.string(),
  value: z.unknown().optional(),
});

export const iPresetStepLoginNexusSchema = iPresetStepBaseSchema.extend({
  type: z.literal("login_nexus"),
});

export const iPresetStepCommandLineSchema = iPresetStepBaseSchema.extend({
  type: z.literal("commandline"),
  arguments: z.array(iCommandLineArgSchema),
});

export const iPresetStepHydrateStateSchema = iPresetStepBaseSchema.extend({
  type: z.literal("hydrate"),
  state: z.unknown(),
});

export const iPresetStepSetGameSchema = iPresetStepBaseSchema.extend({
  type: z.literal("setgame"),
  game: z.string(),
});

export const iPresetStepInstallModSchema = iPresetStepBaseSchema.extend({
  type: z.literal("installmod"),
  url: z.string(),
});

export const iPresetStepRestartSchema = iPresetStepBaseSchema.extend({
  type: z.literal("restart"),
});

export const iPresetStepSchema = z.union([
  iPresetStepCommandLineSchema,
  iPresetStepInstallModSchema,
  iPresetStepHydrateStateSchema,
  iPresetStepSetGameSchema,
  iPresetStepRestartSchema,
  iPresetStepLoginNexusSchema,
]);

export const iPresetStateSchema = z.object({
  completed: z.array(z.string()),
  data: z.unknown(),
});

export const iPresetSchema = z.object({
  steps: z.array(iPresetStepSchema),
});

export const iPresetsStateSchema = z.object({
  presets: z.record(z.string(), iPresetStateSchema),
  processing: z.string().optional(),
});